feat(sources): adds source validate command. uses manifest schema file for structural validation
