+++ crates/coral-spec/src/schema/source_manifest.schema.json
@@ -122,7 +122,7 @@
     "request": {
       "type": "object",
       "additionalProperties": false,
-      "required": ["method", "path"],
+      "required": ["path"],
       "properties": {
         "method": { "type": "string", "enum": ["GET", "POST"] },
         "path": { "type": "string", "minLength": 1 },
