+++ crates/coral-spec/src/validate.rs
@@ -9,25 +9,11 @@ use crate::common::{
 use crate::{ManifestError, ParsedTemplate, Result, TemplateNamespace};
 
 pub(crate) fn validate_manifest_top_level(
-    dsl_version: u32,
     name: &str,
     schema: &str,
     backend: SourceBackend,
     base_url: &ParsedTemplate,
-    table_count: usize,
 ) -> Result<()> {
-    if dsl_version != 3 {
-        return Err(ManifestError::validation(format!(
-            "source '{name}' uses unsupported dsl_version={dsl_version} (expected 3)"
-        )));
-    }
-
-    if table_count == 0 {
-        return Err(ManifestError::validation(format!(
-            "source '{name}' has no tables"
-        )));
-    }
-
     match backend {
         SourceBackend::Http => {
             if base_url.raw().trim().is_empty() {
@@ -59,12 +45,7 @@ pub(crate) fn validate_http_table(
     requests: &[RequestRouteSpec],
     pagination: &PaginationSpec,
 ) -> Result<()> {
-    if request.path.raw().trim().is_empty() {
-        return Err(ManifestError::validation(format!(
-            "{schema}.{table_name} has an empty request.path"
-        )));
-    }
-
+    // Empty request.path is rejected by the JSON Schema (minLength: 1).
     validate_columns(columns, schema, table_name)?;
     let known_filters = validate_filters_and_column_exprs(filters, columns, schema, table_name)?;
 
@@ -140,11 +121,7 @@ pub(crate) fn validate_filters_and_column_exprs(
 ) -> Result<HashSet<String>> {
     let mut known_filters = HashSet::new();
     for filter in filters {
-        if filter.name.trim().is_empty() {
-            return Err(ManifestError::validation(format!(
-                "{schema}.{table} has a filter with an empty name"
-            )));
-        }
+        // Empty filter names are rejected by the JSON Schema (minLength: 1).
         if !known_filters.insert(filter.name.clone()) {
             return Err(ManifestError::validation(format!(
                 "{schema}.{table} has duplicate filter '{}'",
